backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend infrastructure for heckler"
readme = "../README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "MemePy>=1.2.0",
    "pillow>=11.0.0",
    "python-osc>=1.9.3",
    "pyyaml>=6.0.2",
    "websockets>=15.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.4",
]
